first stable version of nodeHtmlLabel
