+++ dash_cytoscape/package.json
@@ -41,6 +41,7 @@
     "cytoscape-euler": "^1.2.1",
     "cytoscape-fcose": "^2.1.0",
     "cytoscape-klay": "^3.1.2",
+    "cytoscape-node-html-label": "^1.2.2",
     "cytoscape-popper": "^2.0.0",
     "cytoscape-spread": "^3.0.0",
     "cytoscape-svg": "0.2.0",
