+++ src/lib/extra_index.js
@@ -15,7 +15,7 @@ import popper from 'cytoscape-popper';
 import './popper.css';
 import UndoRedoList from './components/UndoRedoList.react';
 import '../../node_modules/bootstrap/dist/css/bootstrap.min.css';
-import cyNodeHtmlLabel from './cytoscape-node-html-label.js';
+import cyNodeHtmlLabel from './cyNodeHtmlLabel.js';
 
 CytoscapeJS.use(coseBilkent);
 CytoscapeJS.use(fcose);
